Separated new discoveryScreen components in a component folder
