+++ app/src/androidTest/java/com/lastaoutdoor/lasta/ui/screen/discovery/DiscoveryScreenTest.kt
@@ -11,6 +11,7 @@ import androidx.test.espresso.action.ViewActions
 import com.lastaoutdoor.lasta.data.api.FakeOutdoorActivityRepository
 import com.lastaoutdoor.lasta.di.AppModule
 import com.lastaoutdoor.lasta.ui.MainActivity
+import com.lastaoutdoor.lasta.ui.screen.discovery.components.RangeSearchComposable
 import com.lastaoutdoor.lasta.viewmodel.DiscoveryScreenType
 import com.lastaoutdoor.lasta.viewmodel.DiscoveryScreenViewModel
 import dagger.hilt.android.testing.BindValue
